Move some methods from trainer to rl_trainer
