+++ ml-agents/mlagents/trainers/trainer/rl_trainer.py
@@ -1,12 +1,15 @@
 # # Unity ML-Agents Toolkit
 from typing import Dict
 from collections import defaultdict
+import abc
 
 from mlagents.trainers.optimizer.tf_optimizer import TFOptimizer
 from mlagents.trainers.buffer import AgentBuffer
 from mlagents.trainers.trainer import Trainer
 from mlagents.trainers.exception import UnityTrainerException
 from mlagents.trainers.components.reward_signals import RewardSignalResult
+from mlagents_envs.timers import hierarchical_timer
+from mlagents.trainers.agent_processor import AgentManagerQueue
 
 RewardSignalResults = Dict[str, RewardSignalResult]
 
@@ -60,16 +63,48 @@ class RLTrainer(Trainer):  # pylint: disable=abstract-method
                 )
                 rewards[agent_id] = 0
 
-    def clear_update_buffer(self) -> None:
+    def _clear_update_buffer(self) -> None:
         """
         Clear the buffers that have been built up during inference.
         """
         self.update_buffer.reset_agent()
 
+    @abc.abstractmethod
+    def _is_ready_update(self):
+        """
+        Returns whether or not the trainer has enough elements to run update model
+        :return: A boolean corresponding to wether or not update_model() can be run
+        """
+        return False
+
+    @abc.abstractmethod
+    def _update_policy(self):
+        """
+        Uses demonstration_buffer to update model.
+        """
+        pass
+
     def advance(self) -> None:
         """
-        Steps the trainer, taking in trajectories and updates if ready
+        Steps the trainer, taking in trajectories and updates if ready.
         """
-        super().advance()
-        if not self.should_still_train:
-            self.clear_update_buffer()
+        with hierarchical_timer("process_trajectory"):
+            for traj_queue in self.trajectory_queues:
+                # We grab at most the maximum length of the queue.
+                # This ensures that even if the queue is being filled faster than it is
+                # being emptied, the trajectories in the queue are on-policy.
+                for _ in range(traj_queue.maxlen):
+                    try:
+                        t = traj_queue.get_nowait()
+                        self._process_trajectory(t)
+                    except AgentManagerQueue.Empty:
+                        break
+        if self.should_still_train:
+            if self._is_ready_update():
+                with hierarchical_timer("_update_policy"):
+                    self._update_policy()
+                    for q in self.policy_queues:
+                        # Get policies that correspond to the policy queue in question
+                        q.put(self.get_policy(q.behavior_id))
+        else:
+            self._clear_update_buffer()
